# Use the exported clock API instead of localtime.c