--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -86,4 +86,4 @@ Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.1.1
 SystemRequirements: C++11
 Remotes:
-    r-lib/clock#213
+    r-lib/clock