--- a/inst/WORDLIST
+++ b/inst/WORDLIST
@@ -38,7 +38,6 @@ KiB
 knitr
 lbzip
 Lifecycle
-localtime
 mandreyel
 mio
 mtcars