… inst/COPYRIGHTS
… src/localtime.c
… src/localtime.h