@DavisVaughan 2021-04-09
Okay @jimhester, I'm finished. A few notes:

I've tweaked `check_tz()` to materialize the system timezone if the user passes in `""`. Normally clock can handle `""`, but it involves calling back to `Sys.timezone()` from C++. Since the code where that happens will be running in parallel through vroom, we can't do that for fear of crashing R.

I've removed all mentions of `localtime.c`, including the copyright and the fact that R Core was listed as an author.

I've removed `consumeInteger1()` since date doesn't use 0-indexed month and day values (unlike mktime). This has the small side effect of making `amPm_` 1-index based (because of a change in `consumeString()`), but that doesn't seem too bad.

I've added a few tests regarding `tz = ""`.

I've added a few tests making assumptions about ambiguous/nonexistent times explicit. For ambiguous times, we default to always choosing the earliest of the two times. For nonexistent times, we return `NA`.