@jimhester 2021-04-09
yeah this looks good to me, thank you for working on it! I may not merge this until Monday, I want to hunt down why the windows tests are failing in the current vroom master and get the CI green before merging, but it shouldn't require any changes to what you have done.