@DavisVaughan 2021-04-12
FYI clock 0.2.0 is on CRAN so I have adjusted the required version and removed the Remote